+++ changelog.md
@@ -1,2 +1,4 @@
 - Added better audio file name and ID command suggestion provider
 - Save a backup of the metadata file before saving
+- Deduplicate already existing audio files
+- Added deduplicate config option